// src/theme/variants/generated.ts
export const generatedThemes: Record<string, GeneratedThemeEntry> = {
// ...
  "theme-26": {
    "light": {
      "primary": "#00a33d",
      "onPrimary": "#f3faff",
      "primaryContainer": "#d9f1e2",
      "onPrimaryContainer": "#005a22",
      "inversePrimary": "#004118",
      "secondary": "#e0edf8",
      "onSecondary": "#202a32",
      "secondaryContainer": "#f3f8fc",
      "onSecondaryContainer": "#202a32",
      "tertiary": "#28bc5e",
      "onTertiary": "#f3faff",
      "tertiaryContainer": "#d8f3e2",
      "onTertiaryContainer": "#145e2f",
      "error": "#f22a36",
      "onError": "#f3faff",
      "errorContainer": "#fdd9db",
      "onErrorContainer": "#85171e",
      "background": "#f3faff",
      "onBackground": "#09131a",
      "surface": "#ffffff",
      "onSurface": "#09131a",
      "surfaceVariant": "#e0edf8",
      "onSurfaceVariant": "#4c575f",
      "surfaceDim": "#e0edf8",
      "surfaceBright": "#ffffff",
      "surfaceContainerLowest": "#ffffff",
      "surfaceContainerLow": "#ffffff",
      "surfaceContainer": "#e0edf8",
      "surfaceContainerHigh": "#f5f6f6",
      "surfaceContainerHighest": "#ebeced",
      "outline": "#d3e0ea",
      "outlineVariant": "#d3e0ea",
      "shadow": "#000000",
      "scrim": "#000000",
      "inverseSurface": "#09131a",
      "inverseOnSurface": "#f3faff",
      "surfaceTint": "#00a33d",
      "chart1": "#00a33d",
      "chart2": "#0083e0",
      "chart3": "#f46800",
      "chart4": "#a264f6",
      "chart5": "#f83e54"
    },
    "dark": {
      "primary": "#2fbc5b",
      "onPrimary": "#000408",
      "primaryContainer": "#0c2f17",
      "onPrimaryContainer": "#a1e1b5",
      "inversePrimary": "#6dd08c",
      "secondary": "#172128",
      "onSecondary": "#e9f0f5",
      "secondaryContainer": "#090d10",
      "onSecondaryContainer": "#e9f0f5",
      "tertiary": "#32c364",
      "onTertiary": "#000408",
      "tertiaryContainer": "#0e371c",
      "onTertiaryContainer": "#99e1b2",
      "error": "#f22a36",
      "onError": "#f3faff",
      "errorContainer": "#490d10",
      "onErrorContainer": "#f99fa5",
      "background": "#02080e",
      "onBackground": "#e9f0f5",
      "surface": "#09131a",
      "onSurface": "#e9f0f5",
      "surfaceVariant": "#172128",
      "onSurfaceVariant": "#85919a",
      "surfaceDim": "#02080e",
      "surfaceBright": "#09131a",
      "surfaceContainerLowest": "#09131a",
      "surfaceContainerLow": "#09131a",
      "surfaceContainer": "#172128",
      "surfaceContainerHigh": "#162027",
      "surfaceContainerHighest": "#242e34",
      "outline": "#202a32",
      "outlineVariant": "#202a32",
      "shadow": "#000000",
      "scrim": "#000000",
      "inverseSurface": "#e9f0f5",
      "inverseOnSurface": "#02080e",
      "surfaceTint": "#2fbc5b",
      "chart1": "#2fbc5b",
      "chart2": "#319cfc",
      "chart3": "#ff7f16",
      "chart4": "#b478ff",
      "chart5": "#ff5969"
    }
  }
};